refactor(components): move type definitions into types folder | bh | #2385
