+++ packages/components/src/types/index.ts
@@ -1,1 +1,3 @@
+export * from './aria-types';
+export * from './common-types';
 export * from './instance-types';
