+++ packages/components/src/utils/scrollLock.ts
@@ -1,3 +1,3 @@
-export const setScrollLock = (isOpen: boolean) => {
+export const setScrollLock = (isOpen: boolean): void => {
   document.body.style.overflow = isOpen ? 'hidden' : '';
 };
